Make i18n async #2
- VNode: getInnerHtml() does no longer resolve i18n
- VNode: add resolvei18n(). This method asynchronously tries to resolve
i18n in the same manner getInnerHtml() did. If it finds a match, it
stores the 'raw' innerHTML (with the unresolved i18n keys) in
'rawInnerHtml' and registers an on DOM Listener, which calls
setInnerHtml with the now resolved innerHtml. I'm storing the origina
innerHTML in a field, because this way we can offer a way to re-resolve
i18n, if the locale changes. resolvei18n(), called a second time, will
use rawInnerHtml as base for resolving.
- VNode: call resolvei18n() in constructor, redirect on error
- Resolver: Make get() return a Promise
- Router: add hasRouteRegistered()