Make i18n async #2 (#3)
* Make i18n async #2

- VNode: getInnerHtml() does no longer resolve i18n
- VNode: add resolvei18n(). This method asynchronously tries to resolve
i18n in the same manner getInnerHtml() did. If it finds a match, it
stores the 'raw' innerHTML (with the unresolved i18n keys) in
'rawInnerHtml' and registers an on DOM Listener, which calls
setInnerHtml with the now resolved innerHtml. I'm storing the origina
innerHTML in a field, because this way we can offer a way to re-resolve
i18n, if the locale changes. resolvei18n(), called a second time, will
use rawInnerHtml as base for resolving.
- VNode: call resolvei18n() in constructor, redirect on error
- Resolver: Make get() return a Promise
- Router: add hasRouteRegistered()

* Make i18n resolving lazy

- Do not resolve on construct of a node, but rather on getInnerHtml()
- fuzzy matching of locales
- call setInnerHtml() directly
- ony resovle if locale has changed since the last time we resolved

* Package lock

* Change the flow in which i18n is resolved

- First check for strict resolvers
- If none of them resovles, check non strict ones
- If they also not resolve, reject

* Remove reflect(), refactor duplicate code in resovleI18n